ref: saco codigo que no se usaba
-m "Acomodo algunas clases tambien"

+++ src/Server/modelo/PosicionFija.hpp
@@ -9,7 +9,7 @@ class PosicionFija : public Posicion{
 
         PosicionFija(int coordenadaX,int coordenadaY) : Posicion(){
             this->posicionX=(float)coordenadaX;
-            this->posicionY=coordenadaY;
+            this->posicionY=(float)coordenadaY;
         }
 
 };
